project/pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = ""
authors = ["Your Name <yourname@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.14.1"
asyncpg = "0.30.0"
fastapi = "0.115.7"
SQLAlchemy = "2.0.37"
uvicorn = "0.34.0"
bcrypt = "4.2.1"
python-jose = "3.3.0"
pydantic = { version = "2.10", extras = ["email"] }
passlib = "1.7.4"
python-multipart = "0.0.20"
minio = "7.2.15"
boto3 = "^1.37.0"
botocore = "^1.37.0"
requests = "2.32.3"
pandas = "2.3.0"
delta-spark="4.0.0"
pyspark="4.0.0"
pyarrow="20.0.0"
fsspec="2025.5.1"
s3fs="2025.5.1"
redis="6.2.0"
httpx="0.28.1"
aiotrino = "^0.3.0"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"